+++ projects/datatypes-owc-json/src/public_api.ts
@@ -1,6 +1,0 @@
-/*
- * Public API Surface of owc-json
- */
-
-export * from './lib/owc-json';
-export * from './lib/eoc-owc-json';