+++ projects/demo-maps/src/app/route-components/route-landingpage/route-landingpage.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-route-landingpage',
+  templateUrl: './route-landingpage.component.html',
+  styleUrls: ['./route-landingpage.component.css']
+})
+export class RouteLandingpageComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}