+++ projects/demo-user-info/src/app/route-components/route-user/route-user.component.ts
@@ -1,0 +1,18 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'ukis-route-user',
+  templateUrl: './route-user.component.html',
+  styleUrls: ['./route-user.component.scss'],
+  host: {
+    "[class.content-container]": "true"
+  }
+})
+export class RouteUserComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}