+++ projects/layer-control/src/lib/layerentry/layerentry.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 
 //imports only for typings...
-import { LayerGroup, Layer } from "@ukis/datatypes-layers";
+import { LayerGroup, Layer } from "@ukis/services-layers";
 import { MapStateService } from '@ukis/services-map-state';
 import { LayersService } from '@ukis/services-layers';