+++ projects/services-layers/src/public_api.ts
@@ -1,6 +1,8 @@
 /*
  * Public API Surface of layers
  */
+export * from './lib/types/Layers';
+export * from './lib/types/LayerGroup';
 
 export * from './lib/layers.service';
 export * from './lib/layers.module';