+++ projects/services-owc-json/src/public_api.ts
@@ -1,6 +1,8 @@
 /*
  * Public API Surface of owc-json
  */
+export * from './lib/types/owc-json';
+export * from './lib/types/eoc-owc-json';
 
 export * from './lib/owc-json.service';
 export * from './lib/owc-json.module';